requirements.txt
logbook
discord.py[voice]
kyokai>=1.3.4